+++ lib/haml_lint/linter/unnecessary_interpolation.rb
@@ -1,0 +1,29 @@
+module HamlLint
+  # Checks for unnecessary uses of string interpolation.
+  #
+  # For example, the following two code snippets are equivalent, but the latter
+  # is more concise (and thus preferred):
+  #
+  #   %tag #{expression}
+  #   %tag= expression
+  class Linter::UnnecessaryInterpolation < Linter
+    include LinterRegistry
+
+    def visit_tag(node)
+      inline_content = node.value[:value]
+      return unless inline_content
+
+      if contains_interpolation?(inline_content) &&
+         only_interpolation?(inline_content)
+        add_lint(node, '`%... \#{expression}` can be written without ' \
+                       'interpolation as `%...= expression`')
+      end
+    end
+
+  private
+
+    def only_interpolation?(content)
+      content.lstrip.start_with?('"#{')
+    end
+  end
+end